Added Day Nineteen, Parts One and Two solutions in Python
Sort of. Part Two is not programmatic unless you have a lot of processor
power and/or time. But there's a crude sort of interface with the
show_last function and pdb.set_trace that makes it possible to inspect
the pattern of behavior in the registers, and "fast-forward" the system
so that it doesn't take the eternity that it otherwise would take.

The solution ended up being the sum of all of the divisors of 10551424,
which register 2 gets set to early on and never changes from.